feat(inputs): pull out standalone results component
